🎨 Update my travels's  folder structure
