+++ src/widgets/index.tsx
@@ -1,3 +1,3 @@
 export { default as Header } from './header/ui/Header';
-export { default as UpcomingPlans } from './my-travels/ui/UpcomingPlans';
-export { default as MyPlans } from './my-travels/ui/MyPlans';
+export { default as UpcomingPlans } from './upcomingPlans/ui/UpcomingPlans';
+export { default as MyPlans } from './myPlans/ui/MyPlan';
