+++ src/widgets/myPlans/types/myPlans.ts
@@ -1,14 +1,3 @@
-export interface UpcomingPlanT {
-  plan_id: number;
-  plan_name: string;
-  user_id: number;
-  theme: string;
-  start_date: string;
-  end_date: string;
-  plan_budget: number;
-  number_of_members: number;
-}
-
 interface PlaceT {
   place_id: number;
   place_name: string;
