+++ src/widgets/upcomingPlans/types/upcomingplan.ts
@@ -1,0 +1,10 @@
+export interface UpcomingPlanT {
+  plan_id: number;
+  plan_name: string;
+  user_id: number;
+  theme: string;
+  start_date: string;
+  end_date: string;
+  plan_budget: number;
+  number_of_members: number;
+}
